Rename class LogPathPlatformService to LocalPathService.
