+++ Covid19Radar/Covid19Radar.Android/MainApplication.cs
@@ -82,7 +82,7 @@ namespace Covid19Radar.Droid
         {
             // Services
             container.Register<IBackupAttributeService, BackupAttributeService>(Reuse.Singleton);
-            container.Register<ILogPathPlatformService, LogPathPlatformService>(Reuse.Singleton);
+            container.Register<ILocalPathService, LocalPathService>(Reuse.Singleton);
             container.Register<ILogPeriodicDeleteService, LogPeriodicDeleteService>(Reuse.Singleton);
             container.Register<ISecureStorageDependencyService, Services.SecureStorageService>(Reuse.Singleton);
             container.Register<IPreferencesService, PreferencesService>(Reuse.Singleton);
