+++ Covid19Radar/Covid19Radar.Android/Services/LocalPathService.cs
@@ -5,11 +5,11 @@
 using System;
 using System.IO;
 using Android.Content;
-using Covid19Radar.Services.Logs;
+using Covid19Radar.Services;
 
-namespace Covid19Radar.Droid.Services.Logs
+namespace Covid19Radar.Droid.Services
 {
-    public class LogPathPlatformService : ILogPathPlatformService
+    public class LocalPathService : ILocalPathService
     {
         #region Instance Properties
 
@@ -33,7 +33,7 @@ namespace Covid19Radar.Droid.Services.Logs
 
         #region Constructors
 
-        public LogPathPlatformService()
+        public LocalPathService()
         {
         }
