+++ Covid19Radar/Covid19Radar.iOS/AppDelegate.cs
@@ -185,7 +185,7 @@ namespace Covid19Radar.iOS
         {
             // Services
             container.Register<IBackupAttributeService, BackupAttributeService>(Reuse.Singleton);
-            container.Register<ILogPathPlatformService, LogPathPlatformService>(Reuse.Singleton);
+            container.Register<ILocalPathService, LocalPathService>(Reuse.Singleton);
             container.Register<ILogPeriodicDeleteService, LogPeriodicDeleteService>(Reuse.Singleton);
             container.Register<ISecureStorageDependencyService, Services.SecureStorageService>(Reuse.Singleton);
             container.Register<IPreferencesService, PreferencesService>(Reuse.Singleton);
