+++ Covid19Radar/Covid19Radar.iOS/Covid19Radar.iOS.csproj
@@ -277,7 +277,6 @@
     <Compile Include="Extensions\AlignmentExtensions.cs" />
     <Compile Include="Extensions\Extensions.cs" />
     <Compile Include="Extensions\FormattedStringExtensions.cs" />
-    <Compile Include="Services\Logs\LogPathPlatformService.cs" />
     <Compile Include="Services\Logs\LogPeriodicDeleteService.cs" />
     <Compile Include="Renderers\CustomDatePickerRenderer.cs" />
     <Compile Include="Services\PreferencesService.cs" />
@@ -291,6 +290,7 @@
     <Compile Include="Services\BackupAttributeService.cs" />
     <Compile Include="Services\ExternalNavigationService.cs" />
     <None Include="EntitlementsDebug.plist" />
+    <Compile Include="Services\LocalPathService.cs" />
   </ItemGroup>
   <ItemGroup>
     <ImageAsset Include="Assets.xcassets\Contents.json">
