+++ Covid19Radar/Covid19Radar/Services/ILocalPathService.cs
@@ -2,9 +2,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-namespace Covid19Radar.Services.Logs
+namespace Covid19Radar.Services
 {
-    public interface ILogPathPlatformService
+    public interface ILocalPathService
     {
         string LogsDirPath { get; }
         string LogUploadingTmpPath { get; }
